TBH:  Thanks to a lot of hard work by Todd Hutchison, WRF can now read and TBH:  write both grib1 and grib2 file formats. TBH: TBH:  The new commit contains several new directories, so please use TBH:  "cvs update -d" when updating your working copies to make sure you pick TBH:  up the new directories.  Otherwise, when you commit and tag the new TBH:  directories will not be tagged -- causing lots of confusion later. TBH: TBH:  As before, all source code needed for grib1 I/O is included in the WRF TBH:  source tree (in external/io_grib1 and external/io_grib_share) and grib1 TBH:  I/O is built automatically.  The new grib1 package only requires TBH:  gribmap.txt when grib1 I/O is actually used -- an improvement for TBH:  those who do not use grib1. TBH: TBH:  The grib2 I/O package is not built automatically because it requires an TBH:  external library called "jasper".  If jasper is installed and you wish TBH:  to use grib2, please set environment variables JASPERLIB and JASPERINC TBH:  prior to running the configure program to trigger the build of the TBH:  grib2 I/O package.  The jasper library and the new environment TBH:  variables are described in file external/io_grib2/README.io_grib2. TBH: TBH:  I have tested the new code extensively to complement Todd's testing. TBH:  Test results can be found in file external/io_grib_share/TEST_RESULTS. TBH:  Note that there is one known issue:  RUCLSM initialization chokes when TBH:  grib1 input is used.  This issue will be addressed in a future commit.
git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@1812 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ external/io_grib1/MEL_grib1/gribputbds.c
@@ -175,7 +175,7 @@ int  n, stat=1;
                     &(pBDS_Head_Input->usBit_pack_num),
                     &(pBDS_Head_Input->fPack_null),
                     pfData_Array, 
-		    (unsigned int **) &pvbstr, 
+		    (unsigned long **) &pvbstr, 
 		    sDec_sc_fctr, &lBDS_length, errmsg)) )
           {
           DPRINT2 ("%s:  Pack Spatial returned err=%d\n", func, n);
